django/requirements.txt
Cython==0.20.1
Django==1.6
ExtensionClass==4.1a1
Missing==3.0
MySQL-python==1.2.5
Pillow==2.3.0
RandomWords==0.1.12
South==0.8.4
argparse==1.2.1
beautifulsoup4==4.3.2
django-blog-zinnia==0.14.1
django-classy-tags==0.5.1
django-cors-headers==0.12
django-filter==0.7
django-localflavor==1.0
django-missing==0.1.11
django-model-utils==2.0.3
django-mptt==0.6.0
django-nose==1.2
django-notifications-hq==0.6.2
django-oauth2-provider==0.2.6
django-relationships==0.3.3
django-rest-swagger==0.1.13
django-sekizai==0.7
django-tagging==0.3.2
django-websocket-redis==0.2.3
django-xmlrpc==0.1.5
djangorestframework==2.3.12
factory-boy==2.3.1
gevent==1.0
greenlet==0.4.2
nodeenv==0.7.2
nose==1.3.0
oauth==1.0.1
pyactiveresource==2.0.0
pyparsing==2.0.2
python-dateutil==2.2
pytz==2014.1
redis==2.9.1
requests==2.2.1
shortuuid==0.4
simple-translation==0.8.6
six==1.6.1
stripe==1.14.0
uWSGI==2.0.1
ujson==1.33
websocket-client==0.12.0
wsgiref==0.1.2
